use 'httpx' for GET and POST instead of 'requests' library
queries to the moonscan.io API started failing. Using Wireshark I
tracked it down to use of the "requests" python package which doesn't
support HTTPS2. It was submitting the moonscan.io GET request with
"HTTP/1.1" and then moonscan.io responds with "HTTP/1.1 403 Forbidden"
and the cloudflare error message. Whereas if I hand-crafted the same
GET request as a URL string, it would be accepted and I received JSON
back with the transaction history. Wireshark showed that these
handcrafted URLs were being submitted as HTTP2.

The solution was to migrate GET and POST HTTP messages to use the
"httpx" package instead of "requests". (and with pip you must install
the httpx optional support for http2)

I went ahead and changed subscan.io calls to use httpx also, but with
minimal testing.