decode dex trace event logs
Previously, we were able to decode the input data to a contract
interaction. But in a DEX exchange, there could be slippage so the
actual amount swapped might be slightly different. Therefore, we need
to get the actual event logs to see the exact quantities exchanged.

This commit uses @yifei_huang's script to decode a transaction log file.
Some testing remains, to figure out to handle multiple input values to
a swap.

contributes to #8